+++ sdk/openai/azure-ai-openai-assistants/src/test/java/com/azure/ai/openai/assistants/AzureFilesSyncTest.java
@@ -3,18 +3,19 @@
 
 package com.azure.ai.openai.assistants;
 
+import com.azure.ai.openai.assistants.implementation.models.FileListResponse;
 import com.azure.ai.openai.assistants.models.FileDeletionStatus;
 import com.azure.ai.openai.assistants.models.FilePurpose;
 import com.azure.ai.openai.assistants.models.OpenAIFile;
-import com.azure.ai.openai.assistants.models.PageableList;
 import com.azure.core.http.HttpClient;
 import com.azure.core.http.rest.RequestOptions;
 import com.azure.core.http.rest.Response;
 import com.azure.core.util.BinaryData;
-import com.azure.core.util.serializer.TypeReference;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
+
+import java.util.List;
 
 import static com.azure.ai.openai.assistants.TestUtils.DISPLAY_NAME_WITH_ARGUMENTS;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -40,8 +41,8 @@ public class AzureFilesSyncTest extends AssistantsClientTestBase {
             assertFileEquals(file, fileFromBackend);
 
             // Get file by purpose
-            PageableList<OpenAIFile> files = client.listFiles(filePurpose);
-            assertTrue(files.getData().stream().anyMatch(f -> f.getId().equals(file.getId())));
+            List<OpenAIFile> files = client.listFiles(filePurpose);
+            assertTrue(files.stream().anyMatch(f -> f.getId().equals(file.getId())));
 
             // Delete file
             FileDeletionStatus deletionStatus = client.deleteFile(file.getId());
@@ -65,8 +66,8 @@ public class AzureFilesSyncTest extends AssistantsClientTestBase {
             assertFileEquals(file, fileFromBackend);
 
             // Get file by purpose
-            PageableList<OpenAIFile> files = client.listFiles(filePurpose);
-            assertTrue(files.getData().stream().anyMatch(f -> f.getId().equals(file.getId())));
+            List<OpenAIFile> files = client.listFiles(filePurpose);
+            assertTrue(files.stream().anyMatch(f -> f.getId().equals(file.getId())));
 
             // Delete file
             FileDeletionStatus deletionStatus = client.deleteFile(file.getId());
@@ -91,8 +92,8 @@ public class AzureFilesSyncTest extends AssistantsClientTestBase {
             assertFileEquals(file, fileFromBackend);
 
             // Get file by purpose
-            PageableList<OpenAIFile> files = client.listFiles(filePurpose);
-            assertTrue(files.getData().stream().anyMatch(f -> f.getId().equals(file.getId())));
+            List<OpenAIFile> files = client.listFiles(filePurpose);
+            assertTrue(files.stream().anyMatch(f -> f.getId().equals(file.getId())));
 
             // Delete file
             FileDeletionStatus deletionStatus = client.deleteFile(file.getId());
@@ -122,9 +123,9 @@ public class AzureFilesSyncTest extends AssistantsClientTestBase {
             requestOptions.addQueryParam("purpose", FilePurpose.ASSISTANTS.toString());
             Response<BinaryData> listFilesResponse = client.listFilesWithResponse(requestOptions);
             assertEquals(200, listFilesResponse.getStatusCode());
-            PageableList<OpenAIFile> files = listFilesResponse.getValue()
-                .toObject(new TypeReference<PageableList<OpenAIFile>>() {});
-            assertTrue(files.getData().stream().anyMatch(f -> f.getId().equals(file.getId())));
+            List<OpenAIFile> files = listFilesResponse.getValue()
+                .toObject(FileListResponse.class).getData();
+            assertTrue(files.stream().anyMatch(f -> f.getId().equals(file.getId())));
 
             // Delete file
             Response<BinaryData> deleteResponse = client.deleteFileWithResponse(file.getId(), new RequestOptions());
@@ -156,9 +157,9 @@ public class AzureFilesSyncTest extends AssistantsClientTestBase {
             requestOptions.addQueryParam("purpose", FilePurpose.ASSISTANTS.toString());
             Response<BinaryData> listFilesResponse = client.listFilesWithResponse(requestOptions);
             assertEquals(200, listFilesResponse.getStatusCode());
-            PageableList<OpenAIFile> files = listFilesResponse.getValue()
-                .toObject(new TypeReference<PageableList<OpenAIFile>>() {});
-            assertTrue(files.getData().stream().anyMatch(f -> f.getId().equals(file.getId())));
+            List<OpenAIFile> files = listFilesResponse.getValue()
+                .toObject(FileListResponse.class).getData();
+            assertTrue(files.stream().anyMatch(f -> f.getId().equals(file.getId())));
 
             // Delete file
             Response<BinaryData> deleteResponse = client.deleteFileWithResponse(file.getId(), new RequestOptions());
@@ -191,9 +192,9 @@ public class AzureFilesSyncTest extends AssistantsClientTestBase {
             requestOptions.addQueryParam("purpose", FilePurpose.FINE_TUNE.toString());
             Response<BinaryData> listFilesResponse = client.listFilesWithResponse(requestOptions);
             assertEquals(200, listFilesResponse.getStatusCode());
-            PageableList<OpenAIFile> files = listFilesResponse.getValue()
-                .toObject(new TypeReference<PageableList<OpenAIFile>>() {});
-            assertTrue(files.getData().stream().anyMatch(f -> f.getId().equals(file.getId())));
+            List<OpenAIFile> files = listFilesResponse.getValue()
+                .toObject(FileListResponse.class).getData();
+            assertTrue(files.stream().anyMatch(f -> f.getId().equals(file.getId())));
 
             // Delete file
             Response<BinaryData> deleteResponse = client.deleteFileWithResponse(file.getId(), new RequestOptions());