+++ sdk/openai/azure-ai-openai-assistants/CHANGELOG.md
@@ -16,6 +16,13 @@
   - `OpenAIPageableListOfThreadRun` 
 
 - We've introduced `PageableList<T>` these classes were used, except for `FileListResponse` where we simply return `List<OpenAIFile>`.
+- If you are using `listFilesWithResponse` and need to manually deserialize the `BinaryData` in the response, you can still use `PageableList<T>` like so:
+```java
+client.listFilesWithResponse(requestOptions)
+    .getValue()
+    .toObject(new TypeReference<PageableList<OpenAIFile>>() {})
+    .getData();
+```
 
 ### Bugs Fixed